add lightining indexer prolog for pypto



# message auto-generated for no-merge-commit merge:
!823 merge new_li_prolog into master

add lightining indexer prolog for pypto

Created-by: wangqi_ai
Commit-by: wangqi_ai
Merged-by: turing_project1
Description: add lightining indexer prolog for pypto

See merge request: cann/pypto-dev!823

+++ framework/src/machine/runtime/device_launcher_binding.h
@@ -62,8 +62,8 @@ struct DeviceLauncherConfig {
     DeviceLauncherConfig() = default;
     DeviceLauncherConfig(bool onboard, int tblockdim, int taicpunum)
         : onBoard(onboard), blockdim(tblockdim), aicpuNum(taicpunum) {}
-    DeviceLauncherConfig(int tdynWorkspaceSize) : dynWorkspaceSize(tdynWorkspaceSize) {}
-    DeviceLauncherConfig(int tdynWorkspaceSize, int64_t trepeatNum)
+    DeviceLauncherConfig(int64_t tdynWorkspaceSize) : dynWorkspaceSize(tdynWorkspaceSize) {}
+    DeviceLauncherConfig(int64_t tdynWorkspaceSize, int64_t trepeatNum)
         : dynWorkspaceSize(tdynWorkspaceSize), repeatNum(trepeatNum) {}
     DeviceLauncherConfig(const std::vector<std::uint64_t> &addrs) : hcclContext(addrs) {}
 };
